make index on separate page work again (Issue #116)
For some reason that I don't fully understand yet the output
format string passed to the multipage output capability check
method gets converted to uppercase.

So the check for lowercase "pdf" failed, and none of the file
formats could be identified as supporting multiple pages.

For now it compares for both lower case "pdf" and uppder case
"PDF", and with that the index output on an extra page for
PDF format output works again.